requirements.txt
Flask
matplotlib
psycopg2-binary
python-dotenv